apps/api/requirements.txt
fastapi==0.104.1
httpx==0.25.2
python-multipart==0.0.5
python3-saml==1.16.0
python-jose[cryptography]==3.3.0
motor==3.3.2
pydantic[email]==2.5.2
aiosendgrid==0.1.0
sendgrid==6.11.0
